% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xl-style.R
\name{createHeaderStyles}
\alias{createHeaderStyles}
\title{Create (Multiple) Header Styles}
\usage{
createHeaderStyles(
  n = 1,
  fgFills = NULL,
  halign = "center",
  valign = "center",
  textDecoration = "Bold",
  border = "TopBottomLeftRight",
  ...
)
}
\arguments{
\item{n}{Number of header styles to generate}

\item{fgFills}{foreground fills. If \code{NULL}, generate automatically.}

\item{halign}{Horizontal alignment of cell contents}

\item{valign}{A name Vertical alignment of cell contents}

\item{textDecoration}{Text styling.}

\item{border}{Cell border. A vector of "top", "bottom", "left", "right" or a single string).}

\item{...}{args passed to \code{openxlsx::createStyle}}
}
\value{
A list of S4 \code{Style} object
}
\description{
A wrapper around \code{openxlsx::createStyle()}.
Each arguments of this functions automatically recycled.
}
