% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xl-loc.R
\name{get_startRow_stack}
\alias{get_startRow_stack}
\title{Get Start Rows to Write Stacked Data Frames}
\usage{
get_startRow_stack(ls_df, startRow = 1, gapRow = 0)
}
\arguments{
\item{ls_df}{A list of data.frame}

\item{startRow}{first row to write of a first data.frame}

\item{gapRow}{empty rows (gap) between each data.frame}
}
\value{
A named numeric vector indicating start rows of each data.frame
}
\description{
Get vector of rows to write a stacked (top to bottom) data frame in a single worksheet
}
