% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lbdoc-package.R
\docType{package}
\name{lbdoc-package}
\alias{lbdoc}
\alias{lbdoc-package}
\title{lbdoc: My Package for Office Document}
\description{
This package contain custom functions that manipulate office document files such as Excel, MS Word, MS Powerpoint, etc.
}
\author{
\strong{Maintainer}: Kittipos Sirivongrungson \email{ki11ip0.s.a.s@gmail.com}

}
\keyword{internal}
