% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{rep_args_len}
\alias{rep_args_len}
\title{Replicate dynamic dots args to Specified Length}
\usage{
rep_args_len(..., length.out = 1)
}
\arguments{
\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> arguments to replicate each by \code{length.out}}

\item{length.out}{amount of replication}
}
\value{
a list
}
\description{
Replicate dynamic dots args to Specified Length
}
