% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{rep_list_elements}
\alias{rep_list_elements}
\title{Replicate Elements of List to Specified Length}
\usage{
rep_list_elements(ls = list(), length.out = 1)
}
\arguments{
\item{ls}{list that have elements to replicated}

\item{length.out}{amount of replication}
}
\value{
a list with equal length of elements
}
\description{
Replicate Elements of List to Specified Length
}
